+++ Controller/Adminhtml/Order/POD.php
@@ -15,6 +15,7 @@ use Magento\Backend\App\Action\Context;
 use Magento\Framework\Controller\Result\JsonFactory;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Sales\Api\OrderRepositoryInterface;
+use Magento\Framework\Serialize\Serializer\Json;
 
 class POD extends Action
 {
@@ -45,10 +46,18 @@ class POD extends Action
     protected OrderRepositoryInterface $orderRepository;
 
     /**
+     * @var Json $json
+     */
+    protected Json $json;
+
+    /**
      * @param OrderShipmentRepository $orderShipmentRepository
+     * @param OrderRepositoryInterface $orderRepository
      * @param JsonFactory $jsonFactory
      * @param Context $context
      * @param Data $helper
+     * @param Uber $uber
+     * @param Json $json
      */
     public function __construct(
         OrderShipmentRepository $orderShipmentRepository,
@@ -56,13 +65,15 @@ class POD extends Action
         JsonFactory $jsonFactory,
         Context $context,
         Data $helper,
-        Uber $uber
+        Uber $uber,
+        Json $json
     ) {
         $this->orderShipmentRepository = $orderShipmentRepository;
         $this->orderRepository = $orderRepository;
         $this->jsonFactory = $jsonFactory;
         $this->helper = $helper;
         $this->uber = $uber;
+        $this->json = $json;
         parent::__construct($context);
     }
 
@@ -93,16 +104,39 @@ class POD extends Action
                  * Get Proof of Delivery from API
                  */
                 $orderData = $this->orderRepository->get($orderShipment->getOrderId());
-                $customerId = $this->helper->getCustomerId($orderData->getStoreId());
-                $podData = $this->uber->getProofOfDelivery(
-                    $orderShipment->getUberShippingId(),
-                    $customerId,
-                    $orderData->getStoreId()
-                );
-                return $resultJson->setData([
-                    'origin' => 'api',
-                    'document' => $podData->document ?: ''
-                ]);
+                $verificationData = $orderShipment->getVerification();
+                if($verificationData === null){
+                    $customerId = $this->helper->getCustomerId($orderData->getStoreId());
+                    $podData = $this->uber->getProofOfDelivery(
+                        $orderShipment->getUberShippingId(),
+                        $customerId,
+                        $orderData->getStoreId()
+                    );
+
+                    // Save Verification Data
+                    if($podData['document'] !== ''){
+                        $verificationData = $podData['document'];
+                        $orderShipment->setVerification($verificationData);
+                        $orderShipment->save();
+                    }
+                }
+
+                /**
+                 * If the "decode" fails, it means that the verification is a base 64 image, so I modify the object
+                 * that we return.
+                 */
+                try {
+                    // Json Data
+                    $verificationObjectData = $this->json->unserialize($verificationData);
+                } catch (\Exception $e) {
+                    // Base 64 Image / API Data
+                    $verificationObjectData = [
+                        'origin' => 'api',
+                        'document' => $verificationData ?: ''
+                    ];
+                }
+
+                return $resultJson->setData($verificationObjectData);
             }
         } catch (NoSuchEntityException|\Exception $e) {
             return $resultJson->setData(['error' => true, 'msg' => $e->getMessage()]);