+++ Model/CreateShipment.php
@@ -25,6 +25,7 @@ use Magento\Sales\Model\Order\Shipment\TrackFactory;
 use Magento\Sales\Model\Order\ShipmentRepository;
 use Magento\Sales\Model\OrderRepository;
 use Magento\Shipping\Model\ShipmentNotifier;
+use Magento\Framework\Event\ManagerInterface;
 
 class CreateShipment
 {
@@ -98,12 +99,18 @@ class CreateShipment
     protected TrackFactory $trackFactory;
 
     /**
+     * @var ManagerInterface $eventManager
+     */
+    protected ManagerInterface $eventManager;
+
+    /**
      * @param Uber $uber
      * @param Data $helper
      * @param Registry $registry
-     * @param Converter $converter
-     * @param TrackFactory $trackFactory
      * @param TimezoneInterface $timezone
+     * @param Converter $converter
+     * @param ManagerInterface $eventManager
+     * @param TrackFactory $trackFactory
      * @param OrderRepository $orderRepository
      * @param ShipmentNotifier $shipmentNotifier
      * @param TransactionFactory $transactionFactory
@@ -116,9 +123,10 @@ class CreateShipment
         Uber                        $uber,
         Data                        $helper,
         Registry                    $registry,
-        Converter                   $converter,
-        TrackFactory                $trackFactory,
         TimezoneInterface           $timezone,
+        Converter                   $converter,
+        ManagerInterface            $eventManager,
+        TrackFactory                $trackFactory,
         OrderRepository             $orderRepository,
         ShipmentNotifier            $shipmentNotifier,
         TransactionFactory          $transactionFactory,
@@ -132,6 +140,7 @@ class CreateShipment
         $this->registry = $registry;
         $this->timezone = $timezone;
         $this->converter = $converter;
+        $this->eventManager = $eventManager;
         $this->trackFactory = $trackFactory;
         $this->orderRepository = $orderRepository;
         $this->shipmentNotifier = $shipmentNotifier;
@@ -244,14 +253,17 @@ class CreateShipment
                     ];
                 }
 
-                // Sandbox Mode Webhooks
-                if (!$this->helper->getIntegrationMode($order->getStoreId()) &&
-                    $this->helper->isWebhooksEnabled($order->getStoreId())) {
+                // Sandbox Mode
+                if (!$this->helper->getIntegrationMode($order->getStoreId())) {
                     $deliveryAdditionalData['test_specifications'] = [
                         'robo_courier_specification' => [
                             'mode' => 'auto',
                         ],
                     ];
+
+                    // Unset PickUP & Drop-off data
+                    unset($deliveryAdditionalData['pickup_ready_dt'], $deliveryAdditionalData['pickup_deadline_dt'],
+                        $deliveryAdditionalData['dropoff_ready_dt'], $deliveryAdditionalData['dropoff_deadline_dt']);
                 }
 
                 // Prepara Request
@@ -292,6 +304,12 @@ class CreateShipment
                 } catch (Exception $e) {
                     throw new Exception($e->getMessage());
                 }
+
+                // Dispatch Event
+                $this->eventManager->dispatch('uber_shipment_create', [
+                    'order' => $order,
+                    'shipment' => $uberOrderShipmentRepository
+                ]);
 
                 // Return Response
                 return $uberResponse;
@@ -383,12 +401,12 @@ class CreateShipment
             $orderShipment = $this->trackFactory->create()->getCollection()
                 ->addFieldToFilter('order_id', ['eq' => $order->getEntityId()])
                 ->getFirstItem();
-            $orderShipment->setTrackNumber($trackNumber);
+            $orderShipment->setTrackNumber($trackURL);
             $orderShipment->save();
         } else {
             $shipment->addTrack(
                 $this->trackFactory->create()
-                    ->setNumber($trackNumber)
+                    ->setNumber($trackURL)
                     ->setCarrierCode(self::CARRIER_CODE)
                     ->setTitle($carrierTitle)
             );
@@ -537,7 +555,7 @@ class CreateShipment
         }
         try {
             $orderComment = __('<b>Uber Shipping UUID</b>: %1', $this->formatTrackingNumber($confirmationData['uuid'], true)) . '<br>';
-            $orderComment .= __('<b>Tracking URL</b>: <a href="%1">%1</a>', $confirmationData['tracking_url']) . '<br>';
+            $orderComment .= __('<b>Tracking URL</b>: <a href="%1" target="_blank">%1</a>', $confirmationData['tracking_url']) . '<br>';
             $order->addCommentToStatusHistory(
                 $orderComment,
                 self::DEFAULT_UBER_STATUS
@@ -568,7 +586,7 @@ class CreateShipment
         $area = ($area == 'return') ? 'pickup' : $area;
         $verificationMethod = $this->helper->getVerificationType($storeId, $area);
 
-        // Exclude Identificacion Method
+        // Exclude Identification Method
         if ($excludeIdentificationMethod && $verificationMethod == 'identification') {
             $verificationMethod = 'signature';
         }